Add additional case related to contracts in 'isCapturedWrite'
Now, as we have contracts and know that some functions will call
passed lambdas only in-place and only once, we can relax restrictions
in some case and do not treat such writes as captures write.
Note that write inside in-place called lambda may still be captured,
so we still have to check parents.