Support inline lambdas.
Add support of in-place-called lambdas in compiler:

- New InlinedLocalFunctionDeclarationInstruction is introduced

- The main difference from LocalFunctionDeclarationInstruction is that
it doesn't necessary lead to SINK. More specifically, it depends on
whether control flow can ordinary leave body of
InlinedLocalFunctionDeclarationInstruction (i.e. if SINK is reachable):
if it does, then .next of InlinedFunction is just next instruction and
SINK otherwise.

- To faciliate abovementioned logic, we have to change how reachability
is used. Previously, SINK was always reachable, even in cases when it
wasn't (e.g. in cases with non-local returns or throws, like
'run { return 42 }'). Now, we don't force-add SINK (and ERROR and EXIT
for the sake of consistency) to the list of reachable instructions, so
they *are not always reachable anymore*

- Non-local jumps and labeled returns create some complexity, as SINK
can be unreachable due to local flow, but reachable via jump from some
of the nested lambdas. Thus, we have to make reachability analysis
global (of course, only InlinedFunctions are take part in global
analysis).