Drop effect system caches after each statement
Effect system caches are statement-wide, so we can drop them after each
statement. To do this, we create separate binding trace for each
statement, which will commit to parent everything except effect
system-related slices.

The issue here is that during 'blockLevelVisitor.getTypeInfo' slice
rewrites can occur on LEXICAL_SCOPE slice.
Previously, everything worked because traces passed to
'getBlockReturnedTypeWithWritableScope' were aware of this and
handled this accordingly (e.g., see 'NoScopeRecordCliBindingTrace').
Now, we pass our custom trace to 'blockLevelVisitor.getTypeInfo', so we
have to deal with LEXICAL_SCOPE rewrites on our own.
To do this, we've exposed creation of TemporaryBindingTrace with custom
'allowSliceRewrite' and create trace with allowed rewrites.